logic vs renderer
